+++ source/extensions/filters/http/ext_authz/ext_authz.cc
@@ -5,10 +5,21 @@
 #include "common/http/codes.h"
 #include "common/router/config_impl.h"
 
+#include "extensions/filters/http/well_known_names.h"
+
 namespace Envoy {
 namespace Extensions {
 namespace HttpFilters {
 namespace ExtAuthz {
+
+void FilterConfigPerRoute::merge(const FilterConfigPerRoute& other) {
+  disabled_ = other.disabled_;
+  auto begin_it = other.context_extensions_.begin();
+  auto end_it = other.context_extensions_.end();
+  for (auto it = begin_it; it != end_it; ++it) {
+    context_extensions_[it->first] = it->second;
+  }
+}
 
 void Filter::initiateCall(const Http::HeaderMap& headers) {
   Router::RouteConstSharedPtr route = callbacks_->route();
@@ -20,8 +31,30 @@ void Filter::initiateCall(const Http::HeaderMap& headers) {
     return;
   }
 
-  Filters::Common::ExtAuthz::CheckRequestUtils::createHttpCheck(callbacks_, headers,
-                                                                check_request_);
+  // Fast route - if we are disabled, no need to merge.
+  const FilterConfigPerRoute* specific_per_route_config =
+      Http::Utility::resolveMostSpecificPerFilterConfig<FilterConfigPerRoute>(
+          HttpFilterNames::get().ExtAuthorization, route);
+  if (specific_per_route_config != nullptr) {
+    if (specific_per_route_config->disabled()) {
+      return;
+    }
+  }
+
+  // We are not disabled - get a merged view of the config:
+  auto&& maybe_merged_per_route_config =
+      Http::Utility::getMergedPerFilterConfig<FilterConfigPerRoute>(
+          HttpFilterNames::get().ExtAuthorization, route,
+          [](FilterConfigPerRoute& cfg_base, const FilterConfigPerRoute& cfg) {
+            cfg_base.merge(cfg);
+          });
+
+  Protobuf::Map<ProtobufTypes::String, ProtobufTypes::String> context_extensions;
+  if (maybe_merged_per_route_config) {
+    context_extensions = maybe_merged_per_route_config.value().takeContextExtensions();
+  }
+  Filters::Common::ExtAuthz::CheckRequestUtils::createHttpCheck(
+      callbacks_, headers, std::move(context_extensions), check_request_);
 
   state_ = State::Calling;
   // Don't let the filter chain continue as we are going to invoke check call.